Add serverless-ruby Dockerfile
 ### Why?
This docker image was being hosted in the [Pardner](https://github.com/ShippingEasy/pardner)
repository, which made it difficult to find if you needed to make a
change to the image but didn't know the historical context of where it
was used.

Logically, since it is a public image, it should be in this repository

 ### What?
- Add a `serverless-ruby` directory with a dockerfile for v1.4.0 of the
`serverless-ruby` build image
- Add a README for the serverless-ruby image and version tables for  `1.4.0`
and `1.3.0`.
- Add references to dependencies that rely on this image
- Add `serverless-ruby` to the top level README